Add files via upload
Rename files and bool writeVtk

+++ src/srcCPU/Signal_Calculator.h
@@ -5,6 +5,6 @@
 using namespace std ;
 #include <vector>
 
-vector< vector<double> > Signal_Calculator (vector< vector<double> > locX , vector< vector<double> > locY , vector<double > centX , vector<double > centY ,vector< vector<double> > oldConcentrations , int index ) ;
+vector< vector<double> > Signal_Calculator (vector< vector<double> > locX , vector< vector<double> > locY , vector<double > centX , vector<double > centY ,vector< vector<double> > oldConcentrations , double index ) ;
 
 #endif /* Signal_Calculator_h */
